# internal/controllers/movie/AddInFavorites.go
package movie

import (
	"net/http"
	"strconv"

	models "github.com/Orken1119/Ozinshe/internal/models/movie_models"
	"github.com/gin-gonic/gin"
)

// @Tags		movie
// @Accept		json
// @Produce	json
// @Param        id   path      int  true  "ID"
// @Security	BearerAuth
// @Success	200		{object}	models.SuccessResponse
// @Failure	default	{object}	models.ErrorResponse
// @Router		/movie/add-in-favorits/{id} [post]
func (mr *MovieController) AddInFavorits(c *gin.Context) {
	userID := c.GetUint("userID")

	movieIDStr := c.Param("id")
	movieID, err := strconv.Atoi(movieIDStr)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponse{
			Result: []models.ErrorDetail{
				{
					Code:    "ERROR_FAVORITS_MOVIES",
					Message: "Can't add movie in favorits",
					Metadata: models.Properties{
						Properties1: err.Error(),
					},
				},
			},
		})
		return
	}

	err = mr.MovieRepository.AddInFavorites(c, movieID, int(userID))
	if err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponse{
			Result: []models.ErrorDetail{
				{
					Code:    "ERROR_FAVORITS_MOVIES",
					Message: "Can't add movie in favorits",
					Metadata: models.Properties{
						Properties1: err.Error(),
					},
				},
			},
		})
		return
	}

	c.JSON(http.StatusOK, models.SuccessResponse{Result: "Movie marked as favorite"})
}